fix: cleanup rebase issues
